requirements.txt
django-lazysignup
django
